pyproject.toml
[project]
name = "sandbox-big-data-regz"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "duckdb>=1.4.3",
    "ibis-framework>=11.0.0",
    "jupyter>=1.1.1",
    "narwhals>=2.14.0",
    "numpy>=2.4.0",
    "pandas>=2.3.3",
    "polars>=1.36.0",
    "pyarrow>=22.0.0",
    "pyarrow-hotfix>=0.7",
    "tqdm>=4.67.1",
]
